LP-11431.4: Update parser-generator.js, move generated files into __genfiles__
