+++ package.json
@@ -90,8 +90,9 @@
     "winston": "^3.7.2"
   },
   "scripts": {
-    "build": "rollup -c config/build/rollup.config.js",
-    "build:prodsizecheck": "rollup -c config/build/rollup.config.js --configFormats='esm' --configEnvironment='production'",
+    "gen:parsers": "yarn --cwd packages/kas gen:parsers",
+    "build": "yarn gen:parsers && rollup -c config/build/rollup.config.js",
+    "build:prodsizecheck": "yarn gen:parsers && rollup -c config/build/rollup.config.js --configFormats='esm' --configEnvironment='production'",
     "watch": "rollup -c config/build/rollup.config.js --watch",
     "clean": "rm -rf packages/perseus*/dist && rm -rf packages/perseus*/node_modules",
     "coverage": "yarn run jest --coverage",
